refactor weight chart painting to util
